requirements.txt
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
numpy>=1.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
Pillow>=10.0.0
